BE-835 Fixed sonarcloud code smell issues
